+++ src/core/crypto/index.ts
@@ -13,6 +13,7 @@ export * from "./poseidon";
 export * from "./privateKey";
 export * from "./publicKey";
 export * from "./secp256k1";
+export * from "./secp256r1";
 export * from "./signature";
 export * from "./singleKey";
 export * from "./types";