+++ src/core/crypto/privateKey.ts
@@ -45,6 +45,7 @@ export class PrivateKey {
   public static readonly AIP80_PREFIXES = {
     [PrivateKeyVariants.Ed25519]: "ed25519-priv-",
     [PrivateKeyVariants.Secp256k1]: "secp256k1-priv-",
+    [PrivateKeyVariants.Secp256r1]: "secp256r1-priv-",
   };
 
   /**